+++ src/views/payment-method-view.js
@@ -29,8 +29,6 @@ PaymentMethodView.prototype._initialize = function () {
 
   addSelectionEventHandler(this.element, this._choosePaymentMethod.bind(this));
 
-  html = html.replace(/@DISABLE_MESSAGE/g, this.strings.hasSubscription);
-
   switch (this.paymentMethod.type) {
     case paymentMethodTypes.applePay:
       html = html.replace(/@ICON/g, 'logoApplePay')