+++ src/views/payment-sheet-views/apple-pay-view.js
@@ -8,6 +8,8 @@ var isHTTPS = require('../../lib/is-https');
 var Promise = require('../../lib/promise');
 var paymentOptionIDs = require('../../constants').paymentOptionIDs;
 
+var DEFAULT_APPLE_PAY_SESSION_VERSION = 2;
+
 function ApplePayView() {
   BaseView.apply(this, arguments);
 }
@@ -20,6 +22,9 @@ ApplePayView.prototype.initialize = function () {
   var self = this;
 
   self.applePayConfiguration = assign({}, self.model.merchantConfiguration.applePay);
+  self.applePaySessionVersion = self.applePayConfiguration.applePaySessionVersion || DEFAULT_APPLE_PAY_SESSION_VERSION;
+
+  delete self.applePayConfiguration.applePaySessionVersion;
 
   self.model.asyncDependencyStarting();
 
@@ -55,7 +60,7 @@ ApplePayView.prototype.initialize = function () {
 ApplePayView.prototype._showPaymentSheet = function () {
   var self = this;
   var request = self.applePayInstance.createPaymentRequest(this.applePayConfiguration.paymentRequest);
-  var session = new global.ApplePaySession(2, request);
+  var session = new global.ApplePaySession(self.applePaySessionVersion, request);
 
   session.onvalidatemerchant = function (event) {
     self.applePayInstance.performValidation({
@@ -94,7 +99,10 @@ ApplePayView.prototype.updateConfiguration = function (key, value) {
 ApplePayView.isEnabled = function (options) {
   var gatewayConfiguration = options.client.getConfiguration().gatewayConfiguration;
   var applePayEnabled = gatewayConfiguration.applePayWeb && Boolean(options.merchantConfiguration.applePay);
+  var applePaySessionVersion = options.merchantConfiguration.applePay && options.merchantConfiguration.applePay.applePaySessionVersion;
   var applePayBrowserSupported;
+
+  applePaySessionVersion = applePaySessionVersion || DEFAULT_APPLE_PAY_SESSION_VERSION;
 
   if (!applePayEnabled) {
     return Promise.resolve(false);
@@ -103,6 +111,10 @@ ApplePayView.isEnabled = function (options) {
   applePayBrowserSupported = global.ApplePaySession && isHTTPS.isHTTPS();
 
   if (!applePayBrowserSupported) {
+    return Promise.resolve(false);
+  }
+
+  if (!global.ApplePaySession.supportsVersion(applePaySessionVersion)) {
     return Promise.resolve(false);
   }