+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 unreleased
 ----------
 - Allow `ApplePaySession` version to be set
+- Fix issue where vaulted payment methods have a UI error when deleting them (#474)
 
 1.15.0
 ------