+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 CHANGELOG
 =========
 
-unreleased
-----------
+1.16.0
+------
 - Allow `ApplePaySession` version to be set
 - Fix issue where vaulted payment methods have a UI error when deleting them (#474)
 - Stop halting Drop-in setup when Data Collector fails to load